pyproject.toml
[project]
name = "jasmine"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "dm-pix>=0.4.3",
  "einops>=0.8.0",
  "flax==0.11.2",
  "jax[cuda]==0.7.2",
  "optax>=0.2.3",
  "tyro>=0.8.5",
  "wandb>=0.17.4",
  "grain==0.2.10",
  "array-record>=0.7.2",
  "pre-commit>=4.2.0",
  "pillow>=11.3.0",
  "jaxlpips",
  "imageio[ffmpeg]>=2.37.2",
  "minerl",
  "decord>=0.6.0",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

[tool.uv.sources]
minerl = { git = "https://github.com/minerllabs/minerl" }

[tool.uv]
override-dependencies = [
    "numpy>=1.26.0"
]